requirements.txt
requests
beautifulsoup4
pandas
feedparser
anthropic
python-dotenv
cloudscraper
lxml
sendgrid
supabase==2.*
googlenewsdecoder
aiohttp
